src/test/kotlin/dev/martinclaus/day01/Day1Test.kt
package dev.martinclaus.day01

import dev.martinclaus.readText
import dev.martinclaus.toFilename
import org.junit.jupiter.api.extension.ExtensionContext
import org.junit.jupiter.params.ParameterizedTest
import org.junit.jupiter.params.provider.Arguments
import org.junit.jupiter.params.provider.ArgumentsProvider
import org.junit.jupiter.params.provider.ArgumentsSource
import java.util.stream.Stream
import java.util.stream.Stream.of
import kotlin.test.assertEquals

class Day1Test {
    private val sut = Day1()

    @ParameterizedTest
    @ArgumentsSource(PartIArgumentsProvider::class)
    fun checkPartI(input: String, expected: Long) {
        val result = sut.partI(input)
        assertEquals(expected, result)
    }

    @ParameterizedTest
    @ArgumentsSource(PartIIArgumentsProvider::class)
    fun checkPartII(input: String, expected: Long) {
        val result = sut.partII(input)
        assertEquals(expected, result)
    }

    class PartIArgumentsProvider : ArgumentsProvider {
        override fun provideArguments(context: ExtensionContext): Stream<out Arguments>? = of(
            Arguments.of(
                """
                3   4
                4   3
                2   5
                1   3
                3   9
                3   3
                """.trimIndent(),
                11
            ),
            Arguments.of(1.toFilename().readText(), 2970687)
        )
    }

    class PartIIArgumentsProvider : ArgumentsProvider {
        override fun provideArguments(context: ExtensionContext): Stream<out Arguments>? = of(
            Arguments.of(
                """
                3   4
                4   3
                2   5
                1   3
                3   9
                3   3
                """.trimIndent(),
                31
            ),
            Arguments.of(1.toFilename().readText(), 23963899)
        )
    }
}
